♻️ refactor: 텍스트 스타일 컴포넌트들 위치 변경   #290
