+++ src/components/editor/editor-ui/text-style-selector/text-align-and-color.tsx
